Dostosowanie częście komponentów do Angular Material.

+++ src/app/groups/groups-list/groups-list.component.ts
@@ -1,7 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
+import { MatDialog, MatPaginator, MatTableDataSource } from '@angular/material';
 
 import { GroupsService } from '../groups.service';
+
 import { Group } from '../group';
+import { RemoveGroupDialogComponent } from '../remove-group-dialog/remove-group-dialog.component';
 
 @Component({
   selector: 'app-groups-list',
@@ -10,30 +13,41 @@ import { Group } from '../group';
 })
 export class GroupsListComponent implements OnInit {
 
-  groupsList: Array<Group> = [];
+  dataSource: MatTableDataSource<Group>;
 
-  constructor(private groupsService: GroupsService) {
+  displayedColumns: string[] = ['name', 'description', 'creationDate', 'options'];
+
+  @ViewChild(MatPaginator)
+  paginator: MatPaginator;
+
+  constructor(private groupsService: GroupsService, public dialog: MatDialog) {
     this.groupsService.getGroupList()
       .subscribe((groups: Array<Group>) => {
-        this.groupsList = groups;
+        this.dataSource = new MatTableDataSource<Group>(Array.from(groups));
+        this.dataSource.paginator = this.paginator;
       })
   }
 
   ngOnInit() {
   }
 
-  onOpenModal(id: string): void {
+  onOpenDialog(id: string, name: string): void {
+    const dialogRef = this.dialog.open(RemoveGroupDialogComponent, {
+      width: '500px',
+      data: {id: id, name: name}
+    });
 
-  }
-
-  onRemove(id: string): void {
-    this.groupsService.removeGroup(id)
-      .then(result => {
-        console.log(result);
-      })
-      .catch(error => {
-        console.log(error);
-      })
+    dialogRef.afterClosed().subscribe(id => {
+      if (id) {
+        this.groupsService.removeGroup(id)
+          .then(result => {
+            console.log(result);
+          })
+          .catch(error => {
+            console.log(error);
+          })
+      }
+    });
   }
 
 }
